refactor($compile): 매장주문, 포장주문, 배달주문 Bounded Context 중심의 패키지 구조로 변경
