+++ src/main/java/kitchenpos/orders/eatin/domain/OrderLineItem.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.orders.eatin.domain;
 
 import javax.persistence.*;
 import java.math.BigDecimal;
